LighTDiff/requirements.txt
torch>=1.7
numpy<1.23 
opencv-python
torchvision
scipy
tqdm
basicsr>=1.3.4.0
facexlib>=0.2.0.3
lmdb
pyyaml
tb-nightly
yapf
# mxnet-cu102<1.8.0
# sklearn
easydict
rawpy
kornia==0.5.11
dominate
trimesh
lpips==0.1.4